+++ Module.php
@@ -4,6 +4,7 @@ namespace humhub\modules\themeOrange;
 
 use humhub\modules\ui\view\helpers\ThemeHelper;
 use humhub\models\Setting;
+use humhub\libs\DynamicConfig;
 use Yii;
 use yii\helpers\Url;
 
@@ -28,17 +29,31 @@ class Module extends \humhub\components\Module
 	 * @var string defines the like icon (options: heart, thumbsup, star)
 	 */
 	public $likeIcon = 'heart';
+    
+    const THEME_NAME = "themeOrange";
 	
-	public static function getCommentLinkSetting() {
-		return Yii::$app->getModule('theme-orange')->settings->get('commentLink');
+    public static function getCommentLinkSetting() {
+		$commentLink = Yii::$app->getModule('theme-orange')->settings->get('commentLink');
+        if (empty($commentLink)) {
+            $commentLink = $this->commentLink;
+        }
+        return $commentLink;
 	}
 	
-	public static function getLikeLinkSetting() {
-		return Yii::$app->getModule('theme-orange')->settings->get('likeLink');
+    public static function getLikeLinkSetting() {
+		$likeLink = Yii::$app->getModule('theme-orange')->settings->get('likeLink');
+        if (empty($likeLink)) {
+            $likeLink = $this->likeLink;
+        }
+        return $likeLink;
 	}
 	
 	public static function getLikeIcon() {
-		return Yii::$app->getModule('theme-orange')->settings->get('likeIcon');
+		$likeIcon = Yii::$app->getModule('theme-orange')->settings->get('likeIcon');
+        if (empty($likeIcon)) {
+            $likeIcon = $this->likeIcon;
+        }
+        return $likeIcon;
 	}
 	
 	/*
@@ -54,6 +69,32 @@ class Module extends \humhub\components\Module
 	public function getDescription() {
         return Yii::t('ThemeOrangeModule.base', 'A child theme for HumHub');
     }
+
+    public function enable() {
+    
+        if (parent::enable()) {
+            $this->enableTheme();
+            return true;
+        }
+        return false;
+    }
+    
+    private function enableTheme()
+    {
+        // Already a theme based theme is active
+        foreach (ThemeHelper::getThemeTree(Yii::$app->view->theme) as $theme) {
+            if ($theme->name === self::THEME_NAME) {
+                return;
+            }
+        }
+
+        $theme = ThemeHelper::getThemeByName(self::THEME_NAME);
+        if ($theme !== null) {
+            $theme->activate();
+            DynamicConfig::rewrite();
+        }
+    }
+    
 	/**
 	 * @inheritdoc
 	 */