+++ docs/CHANGELOG.md
@@ -1,3 +1,6 @@
+## TBA
+- Fix: theme was activated when another module was enabled
+
 ## 0.6.0 (March 11 2023)
 - Enh: add configuration page (replace file configuration) - NOTE: Please remove file configuration and select your settings in the new administration page
 - Update less, css and views for HumHub 1.14 (now min version)